HSEARCH-3123 Lucene - Move multitenancy strategies to a separate package
Same reasoning as the previous commit.
