+++ backend-lucene/src/main/java/org/hibernate/search/v6poc/backend/lucene/document/impl/AbstractLuceneDocumentBuilder.java
@@ -13,7 +13,7 @@ import java.util.Objects;
 
 import org.apache.lucene.document.Document;
 import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneIndexSchemaObjectNode;
-import org.hibernate.search.v6poc.backend.lucene.impl.MultiTenancyStrategy;
+import org.hibernate.search.v6poc.backend.lucene.multitenancy.impl.MultiTenancyStrategy;
 import org.hibernate.search.v6poc.logging.impl.Log;
 import org.hibernate.search.v6poc.util.impl.common.LoggerFactory;
