+++ backend-lucene/src/main/java/org/hibernate/search/v6poc/backend/lucene/document/impl/LuceneNestedObjectDocumentBuilder.java
@@ -14,7 +14,7 @@ import org.apache.lucene.document.StringField;
 import org.apache.lucene.index.IndexableField;
 import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneFields;
 import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneIndexSchemaObjectNode;
-import org.hibernate.search.v6poc.backend.lucene.impl.MultiTenancyStrategy;
+import org.hibernate.search.v6poc.backend.lucene.multitenancy.impl.MultiTenancyStrategy;
 
 /**
  * @author Guillaume Smet
