+++ backend-lucene/src/main/java/org/hibernate/search/v6poc/backend/lucene/impl/LuceneBackendFactory.java
@@ -14,6 +14,9 @@ import java.util.Optional;
 import org.hibernate.search.v6poc.backend.lucene.cfg.MultiTenancyStrategyConfiguration;
 import org.hibernate.search.v6poc.backend.lucene.cfg.SearchBackendLuceneSettings;
 import org.hibernate.search.v6poc.backend.lucene.logging.impl.Log;
+import org.hibernate.search.v6poc.backend.lucene.multitenancy.impl.DiscriminatorMultiTenancyStrategyImpl;
+import org.hibernate.search.v6poc.backend.lucene.multitenancy.impl.MultiTenancyStrategy;
+import org.hibernate.search.v6poc.backend.lucene.multitenancy.impl.NoMultiTenancyStrategyImpl;
 import org.hibernate.search.v6poc.backend.lucene.work.impl.StubLuceneWorkFactory;
 import org.hibernate.search.v6poc.backend.spi.BackendImplementor;
 import org.hibernate.search.v6poc.backend.spi.BackendFactory;
