+++ backend-lucene/src/main/java/org/hibernate/search/v6poc/backend/lucene/multitenancy/impl/NoMultiTenancyStrategyImpl.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.v6poc.backend.lucene.impl;
+package org.hibernate.search.v6poc.backend.lucene.multitenancy.impl;
 
 import java.lang.invoke.MethodHandles;
 
@@ -17,7 +17,7 @@ import org.hibernate.search.v6poc.backend.lucene.work.impl.TermBasedUpdateEntryL
 import org.hibernate.search.v6poc.backend.spi.BackendImplementor;
 import org.hibernate.search.v6poc.util.impl.common.LoggerFactory;
 
-class NoMultiTenancyStrategyImpl implements MultiTenancyStrategy {
+public class NoMultiTenancyStrategyImpl implements MultiTenancyStrategy {
 
 	private static final Log log = LoggerFactory.make( Log.class, MethodHandles.lookup() );
